Move consumer factory interface outside connector (#667)
Move consumer factory interface outside connector
