+++ datastream-kafka-connector/src/main/java/com/linkedin/datastream/connectors/kafka/mirrormaker/NoOpTopicManager.java
@@ -16,7 +16,7 @@ import com.linkedin.datastream.common.Datastream;
 import com.linkedin.datastream.common.DatastreamRuntimeException;
 import com.linkedin.datastream.connectors.CommonConnectorMetrics;
 import com.linkedin.datastream.connectors.kafka.GroupIdConstructor;
-import com.linkedin.datastream.connectors.kafka.KafkaConsumerFactory;
+import com.linkedin.datastream.kafka.factory.KafkaConsumerFactory;
 import com.linkedin.datastream.server.DatastreamTask;
