+++ datastream-kafka/src/main/java/com/linkedin/datastream/kafka/factory/KafkaConsumerFactoryImpl.java
@@ -3,7 +3,7 @@
  *  Licensed under the BSD 2-Clause License. See the LICENSE file in the project root for license information.
  *  See the NOTICE file in the project root for additional information regarding copyright ownership.
  */
-package com.linkedin.datastream.connectors.kafka;
+package com.linkedin.datastream.kafka.factory;
 
 import java.util.Properties;
